undo circ import debugging
