+++ example_rules/config/labels.yaml
@@ -2,4 +2,4 @@ labels:
   meow:
     valid_for: [User]
     connotation: positive
-    description: ayu testing label
+    description: testing label
