+++ osprey_worker/src/osprey/worker/lib/storage/postgres.py
@@ -8,13 +8,12 @@ from typing import TYPE_CHECKING, Dict, Iterator, Optional  # noqa: E402
 
 from flask import Flask, has_request_context  # noqa: E402
 from osprey.worker.lib.config import Config  # noqa: E402
+from osprey.worker.lib.singletons import CONFIG  # noqa: E402
 from sqlalchemy import MetaData  # noqa: E402
 from sqlalchemy.engine.url import make_url  # noqa: E402
 from sqlalchemy.ext.declarative import declarative_base  # noqa: E402
 from sqlalchemy.orm import Session, sessionmaker  # noqa: E402
 from sqlalchemy.orm.scoping import ThreadLocalRegistry  # type: ignore # missing stub  # noqa: E402
-
-# from osprey.worker.lib.singletons import CONFIG  # noqa: E402
 
 metadata = MetaData()
 Model = declarative_base(name='Model', metadata=metadata)
@@ -51,9 +50,6 @@ def init_from_config(database: str) -> None:
         Session.configure(bind=new_engine)
         # Create all tables defined in the metadata
         metadata.create_all(new_engine)
-
-    # prevent circular imports
-    from osprey.worker.lib.singletons import CONFIG  # noqa: E402
 
     CONFIG.instance().register_configuration_callback(_init)
